Struct partial_ref::Ref

pub struct Ref<'a, Target: PartialRefTarget + ?Sized> { /* private fields */ }

An empty partial reference borrowing no parts.

Partial references with a non-empty set of borrowed parts are built by nesting this type within the Mut and Const types.