Trait partial_ref::HasSubset

pub unsafe trait HasSubset<'a, Reference, SubsetIndex>: PartialRef<'a> {
    type Remainder: PartialRef<'a, Target = Self::Target>;
}

(internal) Asserts that the parts of the partial reference Reference are a subset of the parts of the partial reference having this trait.

A list of parts is considered a subset if they can be plucked in sequence.

Required Associated Types

type Remainder: PartialRef<'a, Target = Self::Target>

Object Safety

This trait is not object safe.