Trait partial_ref::PartSpec

pub trait PartSpec<Part> { }

Helper trait to strip lifetimes from a part.

Every part SomePart<'a, ...> should implement PartSpec<SomePart<'b, ...>>. This is used by the part, part_mut, split_part and split_part_mut functions. This allows the passed parameter to have a different lifetime than the accessed part. This in turn enables part selection using globals with static lifetimes as declared by the part macro.