Trait partial_ref::PluckConst

pub unsafe trait PluckConst<'a, PluckedPart, Index>: PartialRef<'a> {
    type Remainder: PartialRef<'a, Target = Self::Target>;
}

(internal) Extracts the constant part PluckedPart at position Index from the partial reference having this trait, leaving Self::Remainder.

Plucking a constant part still leaves the part in the remaining reference, but will change it from mutable to constant.

The Index type can be inferred.

Required Associated Types

type Remainder: PartialRef<'a, Target = Self::Target>

The partial reference left after plucking.

Object Safety

This trait is not object safe.